requirements.txt
langchain
openai
langchain-community
langchain-openai
gradio
google-generativeai
langchain-google-genai